+++ src/main/account/AccountList.java
@@ -1,4 +1,6 @@
-package main;
+package main.account;
+
+import main.Role;
 
 import java.util.ArrayList;
