+++ src/main/account/Admin.java
@@ -1,4 +1,7 @@
-package main;
+package main.account;
+
+import main.Role;
+import main.account.Account;
 
 public class Admin extends Account {
     public Admin(int salary, int balance, Enum<Role> role, String username, String password) {
