+++ src/main/account/CreateAccount.java
@@ -1,4 +1,4 @@
-package main;
+package main.account;
 
 public class CreateAccount {
     private static int accountIdCounter = -1;
