+++ src/main/account/User.java
@@ -1,4 +1,7 @@
-package main;
+package main.account;
+
+import main.Role;
+import main.account.Account;
 
 public class User extends Account {
     public User(int salary, int balance, Enum<Role> role, String username, String password) {
